# swagger-output.json
{
  "swagger": "2.0",
  "info": {
    "title": "Freemily",
    "description": "Freemily 백엔드 서버를 nodejs로 재구현한 프로젝트입니다.",
    "version": "1.0.0"
  },
  "host": "localhost:3001",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/users/": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "유저 생성 POST 요청",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "유저 생성 정보",
            "schema": {
              "type": "object",
              "properties": {
                "userName": {
                  "type": "string",
                  "example": "Hello"
                },
                "userNickname": {
                  "type": "string",
                  "example": "Hi"
                },
                "userEmail": {
                  "type": "string",
                  "example": "hehe@naver.com"
                },
                "userPassword": {
                  "type": "string",
                  "example": "mypassword11"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "그룹 생성",
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "type": "number",
                  "example": 1
                },
                "userName": {
                  "type": "string",
                  "example": "Hello"
                },
                "userNickname": {
                  "type": "string",
                  "example": "Hi"
                },
                "userEmail": {
                  "type": "string",
                  "example": "hehe@naver.com"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "잘못된 request 형식",
            "schema": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "array",
                  "example": [
                    "Email format is not valid",
                    "Email is empty",
                    "Password is too short",
                    "Name is too short"
                  ],
                  "items": {
                    "type": "string"
                  }
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "409": {
            "description": "이미 가입한 이메일",
            "schema": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "string",
                  "example": "Email is already use"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/users/login": {
      "post": {
        "tags": [
          "User"
        ],
        "description": "유저 로그인 POST 요청",
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "description": "유저 로그인 정보",
            "schema": {
              "type": "object",
              "properties": {
                "userEmail": {
                  "type": "string",
                  "example": "hehe@naver.com"
                },
                "userPassword": {
                  "type": "string",
                  "example": "mypassword11"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "그룹 생성",
            "schema": {
              "type": "object",
              "properties": {
                "userId": {
                  "type": "number",
                  "example": 1
                },
                "userName": {
                  "type": "string",
                  "example": "Hello"
                },
                "userNickname": {
                  "type": "string",
                  "example": "Hi"
                },
                "userEmail": {
                  "type": "string",
                  "example": "hehe@naver.com"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "400": {
            "description": "잘못된 request 형식",
            "schema": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "string",
                  "example": "[Email format is not valid] or [Password is empty]"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "403": {
            "description": "이메일과 비밀번호가 일치하지 않음",
            "schema": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "string",
                  "example": "Email and password mismatch"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          },
          "404": {
            "description": "이메일을 찾을 수 없거나 그룹에 가입하지 않음",
            "schema": {
              "type": "object",
              "properties": {
                "msg": {
                  "type": "string",
                  "example": "[Email not found] or [User not join group]"
                }
              },
              "xml": {
                "name": "main"
              }
            }
          }
        }
      }
    },
    "/users/{user_id}": {
      "delete": {
        "description": "",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "userPassword": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          }
        }
      },
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "name": {
                  "example": "any"
                },
                "nickname": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
    "/users/{user_id}/password": {
      "put": {
        "description": "",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "userPassword": {
                  "example": "any"
                },
                "userNewPassword": {
                  "example": "any"
                },
                "userNewPasswordCheck": {
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "404": {
            "description": "Not Found"
          },
          "409": {
            "description": "Conflict"
          }
        }
      }
    },
    "/users/details/{user_id}": {
      "get": {
        "description": "",
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "date",
            "in": "query",
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "404": {
            "description": "Not Found"
          }
        }
      }
    },
...
  }
}